Refactored schemas and fixed test and doc
