+++ internal/service/project/data_source_projects.go
@@ -5,11 +5,11 @@ import (
 	"fmt"
 
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
-	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/id"
-	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/constant"
+
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/conversion"
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/config"
 	"go.mongodb.org/atlas-sdk/v20250312004/admin"
@@ -32,154 +32,8 @@ type ProjectsDS struct {
 	config.DSCommon
 }
 
-type tfProjectsDSModel struct {
-	ID           types.String        `tfsdk:"id"`
-	Results      []*TFProjectDSModel `tfsdk:"results"`
-	PageNum      types.Int64         `tfsdk:"page_num"`
-	ItemsPerPage types.Int64         `tfsdk:"items_per_page"`
-	TotalCount   types.Int64         `tfsdk:"total_count"`
-}
-
 func (d *ProjectsDS) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
-	resp.Schema = schema.Schema{
-		Attributes: map[string]schema.Attribute{
-			// https://github.com/hashicorp/terraform-plugin-testing/issues/84#issuecomment-1480006432
-			"id": schema.StringAttribute{ // required by hashicorps terraform plugin testing framework
-				DeprecationMessage: "Please use each project's id attribute instead",
-				Computed:           true,
-			},
-			"page_num": schema.Int64Attribute{
-				Optional: true,
-			},
-			"items_per_page": schema.Int64Attribute{
-				Optional: true,
-			},
-			"total_count": schema.Int64Attribute{
-				Computed: true,
-			},
-			"results": schema.ListNestedAttribute{
-				Computed: true,
-				NestedObject: schema.NestedAttributeObject{
-					Attributes: map[string]schema.Attribute{
-						"id": schema.StringAttribute{
-							Computed: true,
-						},
-						"org_id": schema.StringAttribute{
-							Computed: true,
-						},
-						"project_id": schema.StringAttribute{
-							Computed: true,
-						},
-						"name": schema.StringAttribute{
-							Computed: true,
-						},
-						"cluster_count": schema.Int64Attribute{
-							Computed: true,
-						},
-						"created": schema.StringAttribute{
-							Computed: true,
-						},
-						"is_collect_database_specifics_statistics_enabled": schema.BoolAttribute{
-							Computed: true,
-						},
-						"is_data_explorer_enabled": schema.BoolAttribute{
-							Computed: true,
-						},
-						"is_extended_storage_sizes_enabled": schema.BoolAttribute{
-							Computed: true,
-						},
-						"is_performance_advisor_enabled": schema.BoolAttribute{
-							Computed: true,
-						},
-						"is_realtime_performance_panel_enabled": schema.BoolAttribute{
-							Computed: true,
-						},
-						"is_schema_advisor_enabled": schema.BoolAttribute{
-							Computed: true,
-						},
-						"is_slow_operation_thresholding_enabled": schema.BoolAttribute{
-							DeprecationMessage: constant.DeprecationParam, // added deprecation in CLOUDP-293855 because was deprecated in the doc
-							Computed:           true,
-						},
-						"region_usage_restrictions": schema.StringAttribute{
-							Computed: true,
-						},
-						"teams": schema.ListNestedAttribute{
-							Computed: true,
-							NestedObject: schema.NestedAttributeObject{
-								Attributes: map[string]schema.Attribute{
-									"team_id": schema.StringAttribute{
-										Computed: true,
-									},
-									"role_names": schema.ListAttribute{
-										Computed:    true,
-										ElementType: types.StringType,
-									},
-								},
-							},
-						},
-						"limits": schema.SetNestedAttribute{
-							Computed: true,
-							NestedObject: schema.NestedAttributeObject{
-								Attributes: map[string]schema.Attribute{
-									"name": schema.StringAttribute{
-										Computed: true,
-									},
-									"value": schema.Int64Attribute{
-										Computed: true,
-									},
-									"current_usage": schema.Int64Attribute{
-										Computed: true,
-									},
-									"default_limit": schema.Int64Attribute{
-										Computed: true,
-									},
-									"maximum_limit": schema.Int64Attribute{
-										Computed: true,
-									},
-								},
-							},
-						},
-						"ip_addresses": schema.SingleNestedAttribute{
-							Computed:           true,
-							DeprecationMessage: fmt.Sprintf(constant.DeprecationParamWithReplacement, "mongodbatlas_project_ip_addresses data source"),
-							Attributes: map[string]schema.Attribute{
-								"services": schema.SingleNestedAttribute{
-									Computed: true,
-									Attributes: map[string]schema.Attribute{
-										"clusters": schema.ListNestedAttribute{
-											Computed: true,
-											NestedObject: schema.NestedAttributeObject{
-												Attributes: map[string]schema.Attribute{
-													"cluster_name": schema.StringAttribute{
-														Computed: true,
-													},
-													"inbound": schema.ListAttribute{
-														ElementType: types.StringType,
-														Computed:    true,
-													},
-													"outbound": schema.ListAttribute{
-														ElementType: types.StringType,
-														Computed:    true,
-													},
-												},
-											},
-										},
-									},
-								},
-							},
-						},
-						"tags": schema.MapAttribute{
-							ElementType: types.StringType,
-							Computed:    true,
-						},
-						"users": DSUsersSchema,
-					},
-				},
-			},
-		},
-	}
-	conversion.UpdateSchemaDescription(&resp.Schema)
+	resp.Schema = pluralDataSourceProjectSchema(ctx)
 }
 
 func (d *ProjectsDS) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
